+++ app/src/main/java/io/minoro75/demoweatherapp/domain/city_name_entity/ObservationLocation.kt
@@ -1,0 +1,15 @@
+package io.minoro75.demoweatherapp.domain.city_name_entity
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class ObservationLocation(
+    val country: String,
+    val state: String,
+    val city: String,
+    val latitude: Double,
+    val longitude: Double,
+    val distance: Double,
+    val timezone: Int
+
+)
