+++ app/src/main/java/io/minoro75/demoweatherapp/domain/city_name_entity/Observations.kt
@@ -1,0 +1,8 @@
+package io.minoro75.demoweatherapp.domain.city_name_entity
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class Observations(
+    val location: List<ObservationLocation>
+)
