+++ app/src/main/java/io/minoro75/demoweatherapp/domain/forecast_entity/ForecastRoot.kt
@@ -1,0 +1,10 @@
+package io.minoro75.demoweatherapp.domain.forecast_entity
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class ForecastRoot(
+    var forecasts: Forecasts,
+    var feedCreation: String,
+    var metric: Boolean
+)
