+++ app/src/main/java/io/minoro75/demoweatherapp/remote/RemoteDataSource.kt
@@ -1,0 +1,9 @@
+package io.minoro75.demoweatherapp.remote
+
+import io.minoro75.demoweatherapp.domain.forecast_entity.ForecastLocation
+import kotlinx.coroutines.flow.Flow
+
+interface RemoteDataSource {
+    suspend fun getWeatherInCity(city: String): Flow<ForecastLocation>
+    suspend fun getCityNameFromCoordinates(lat: Double, lon: Double): Flow<String>
+}
