+++ app/src/main/java/io/minoro75/demoweatherapp/ui/MainActivity.kt
@@ -1,0 +1,25 @@
+package io.minoro75.demoweatherapp.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.findNavController
+import by.kirich1409.viewbindingdelegate.viewBinding
+import dagger.hilt.android.AndroidEntryPoint
+import io.minoro75.demoweatherapp.R
+import io.minoro75.demoweatherapp.databinding.ActivityMainBinding
+import timber.log.Timber
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity(R.layout.activity_main) {
+
+    private val binding: ActivityMainBinding by viewBinding()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        Timber.plant()
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onSupportNavigateUp(): Boolean {
+        return findNavController(R.id.nav_host_fragment).navigateUp()
+    }
+}
