+++ app/src/main/java/io/minoro75/demoweatherapp/utils/BaseApp.kt
@@ -1,0 +1,11 @@
+package io.minoro75.demoweatherapp.utils
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class BaseApp : Application() {
+    override fun onCreate() {
+        super.onCreate()
+    }
+}
