+++ app/src/main/java/io/minoro75/demoweatherapp/utils/Extensions.kt
@@ -1,0 +1,11 @@
+package io.minoro75.demoweatherapp.utils
+
+import android.view.View
+
+fun View.toVisible() {
+    visibility = View.VISIBLE
+}
+
+fun View.toInvisible() {
+    visibility = View.INVISIBLE
+}
