+++ app/src/main/java/io/minoro75/demoweatherapp/utils/Status.kt
@@ -1,0 +1,7 @@
+package io.minoro75.demoweatherapp.utils
+
+open class Status {
+    object SUCCESS : Status()
+    object ERROR : Status()
+    object LOADING : Status()
+}
