--- app/src/main/java/io/minoro75/heremapsweatherapp/domain/city_name_entity/CityNameRoot.kt
@@ -1,10 +1,0 @@
-package io.minoro75.heremapsweatherapp.domain.city_name_entity
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class CityNameRoot(
-    val observations: Observations,
-    val feedCreation: String,
-    val metric: Boolean
-)
