--- app/src/main/java/io/minoro75/heremapsweatherapp/domain/city_name_entity/Observations.kt
@@ -1,8 +1,0 @@
-package io.minoro75.heremapsweatherapp.domain.city_name_entity
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Observations(
-    val location: List<ObservationLocation>
-)
