--- app/src/main/java/io/minoro75/heremapsweatherapp/domain/forecast_entity/ForecastLocation.kt
@@ -1,15 +1,0 @@
-package io.minoro75.heremapsweatherapp.domain.forecast_entity
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class ForecastLocation(
-    var forecast: List<Weather>,
-    var country: String,
-    var state: String,
-    var city: String,
-    var latitude: Double,
-    var longitude: Double,
-    var distance: Int,
-    var timezone: Int
-)
