--- app/src/main/java/io/minoro75/heremapsweatherapp/domain/forecast_entity/ForecastRoot.kt
@@ -1,10 +1,0 @@
-package io.minoro75.heremapsweatherapp.domain.forecast_entity
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class ForecastRoot(
-    var forecasts: Forecasts,
-    var feedCreation: String,
-    var metric: Boolean
-)
