--- app/src/main/java/io/minoro75/heremapsweatherapp/domain/forecast_entity/Forecasts.kt
@@ -1,8 +1,0 @@
-package io.minoro75.heremapsweatherapp.domain.forecast_entity
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Forecasts(
-    var forecastLocation: ForecastLocation
-)
