--- app/src/main/java/io/minoro75/heremapsweatherapp/ui/MainActivity.kt
@@ -1,25 +1,0 @@
-package io.minoro75.heremapsweatherapp.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.findNavController
-import by.kirich1409.viewbindingdelegate.viewBinding
-import dagger.hilt.android.AndroidEntryPoint
-import io.minoro75.heremapsweatherapp.R
-import io.minoro75.heremapsweatherapp.databinding.ActivityMainBinding
-import timber.log.Timber
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity(R.layout.activity_main) {
-
-    private val binding: ActivityMainBinding by viewBinding()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        Timber.plant()
-        super.onCreate(savedInstanceState)
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        return findNavController(R.id.nav_host_fragment).navigateUp()
-    }
-}
