--- app/src/main/java/io/minoro75/heremapsweatherapp/utils/BaseApp.kt
@@ -1,11 +1,0 @@
-package io.minoro75.heremapsweatherapp.utils
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class BaseApp : Application() {
-    override fun onCreate() {
-        super.onCreate()
-    }
-}
