--- app/src/main/java/io/minoro75/heremapsweatherapp/utils/Extensions.kt
@@ -1,11 +1,0 @@
-package io.minoro75.heremapsweatherapp.utils
-
-import android.view.View
-
-fun View.toVisible() {
-    visibility = View.VISIBLE
-}
-
-fun View.toInvisible() {
-    visibility = View.INVISIBLE
-}
