--- app/src/main/java/io/minoro75/heremapsweatherapp/utils/Status.kt
@@ -1,7 +1,0 @@
-package io.minoro75.heremapsweatherapp.utils
-
-open class Status {
-    object SUCCESS : Status()
-    object ERROR : Status()
-    object LOADING : Status()
-}
